# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: TransportPro - Transportation Business Management API
servers:
  - url: /api
...
paths:
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List invoices
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice from batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        batchId:
          type: integer
        batchName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        totalLoadedQty:
          type: number
        totalDeliveredQty:
          type: number
        ratePerMt:
          type: number
        grossRevenue:
          type: number
        totalShortCharge:
          type: number
        netRevenue:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        issuedDate:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, invoiceNumber, batchId, batchName, clientId, clientName, totalLoadedQty, totalDeliveredQty, ratePerMt, grossRevenue, totalShortCharge, netRevenue, status, createdAt]
    InvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/Invoice"
        - type: object
          properties:
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/InvoiceLineItem"
          required: [lineItems]
    InvoiceLineItem:
      type: object
      properties:
        tripId:
          type: integer
        truckPlate:
          type: string
        product:
          type: string
        loadedQty:
          type: number
        deliveredQty:
          type: number
        shortQty:
          type: number
        allowanceQty:
          type: number
        chargeableShort:
          type: number
        shortCharge:
          type: number
        grossRevenue:
          type: number
        netRevenue:
          type: number
      required: [tripId, truckPlate, product, loadedQty, deliveredQty, shortQty, allowanceQty, chargeableShort, shortCharge, grossRevenue, netRevenue]
    InvoiceInput:
      type: object
      properties:
        batchId:
          type: integer
        issuedDate:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, paid, overdue]
      required: [batchId, status]
    InvoiceUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        issuedDate:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...